requirements.txt
aiogram==3.11.0
asyncio==3.4.3
python-dotenv==1.0.1
aiohttp==3.10.2
supabase==2.5.3
pytest==8.3.1
pytest-asyncio==0.23.8
APScheduler~=3.10.4
redis~=5.0.7
apscheduler_di~=0.1.0
httpx==0.27.0
pytz==2024.1
